+++ packages/llamaindex/src/embeddings/index.ts
@@ -1,5 +1,0 @@
-export * from "@llamaindex/core/embeddings";
-export { FireworksEmbedding } from "./fireworks.js";
-export * from "./JinaAIEmbedding.js";
-export * from "./OpenAIEmbedding.js";
-export { TogetherEmbedding } from "./together.js";