+++ packages/llamaindex/src/index.edge.ts
@@ -22,6 +22,7 @@ export {
 export * from "@llamaindex/core/agent";
 export * from "@llamaindex/core/chat-engine";
 export * from "@llamaindex/core/data-structs";
+export * from "@llamaindex/core/embeddings";
 export {
   CallbackManager,
   DEFAULT_BASE_URL,
@@ -65,10 +66,10 @@ export * from "@llamaindex/core/storage/doc-store";
 export * from "@llamaindex/core/storage/index-store";
 export * from "@llamaindex/core/storage/kv-store";
 export * from "@llamaindex/core/utils";
+export * from "@llamaindex/openai";
 export * from "@llamaindex/workflow/agent";
 export * from "./agent/index.js";
 export * from "./cloud/index.js";
-export * from "./embeddings/index.js";
 export * from "./engines/chat/index.js";
 export * from "./engines/query/index.js";
 export * from "./evaluation/index.js";
@@ -76,7 +77,6 @@ export * from "./extractors/index.js";
 export * from "./indices/index.js";
 export * from "./ingestion/index.js";
 export { imageToDataUrl } from "./internal/utils.js";
-export * from "./llm/index.js";
 export * from "./node-parser.js";
 export * from "./objects/index.js";
 export * from "./OutputParser.js";