+++ packages/llamaindex/src/index.ts
@@ -1,8 +1,5 @@
 export * from "./index.edge.js";
 
-// TODO: clean up, move to jinaai package
-export { JinaAIEmbedding } from "./embeddings/JinaAIEmbedding.js";
-
 // Don't export file-system stores for non-node.js runtime on top level,
 //  as we cannot guarantee that they will work in other environments
 export * from "./storage/index.js";