+++ e2e/examples/vite-import-llamaindex/CHANGELOG.md
@@ -1,5 +1,12 @@
 # vite-import-llamaindex
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [beb922b]
+  - llamaindex@0.9.7
+
 ## 0.0.6
 
 ### Patch Changes