+++ packages/providers/replicate/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @llamaindex/replicate
 
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [beb922b]
+  - @llamaindex/env@0.1.29
+  - @llamaindex/core@0.5.6
+
 ## 0.0.38
 
 ### Patch Changes