+++ packages/providers/storage/firestore/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @llamaindex/firestore
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [beb922b]
+  - @llamaindex/env@0.1.29
+  - @llamaindex/core@0.5.6
+
 ## 1.0.3
 
 ### Patch Changes