+++ packages/providers/vllm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @llamaindex/vllm
 
+## 0.0.28
+
+### Patch Changes
+
+- @llamaindex/openai@0.1.58
+
 ## 0.0.27
 
 ### Patch Changes