+++ packages/providers/voyage-ai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @llamaindex/voyage-ai
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [beb922b]
+  - @llamaindex/env@0.1.29
+  - @llamaindex/core@0.5.6
+
 ## 1.0.2
 
 ### Patch Changes